Break apart the swap quote styles
