+++ src/components/scenes/CryptoExchangeQuoteProcessingScene.js
@@ -1,11 +1,12 @@
 // @flow
 
 import React from 'react'
-import { ActivityIndicator, View } from 'react-native'
+import { ActivityIndicator, StyleSheet, View } from 'react-native'
 
 import s from '../../locales/strings.js'
 import FormattedText from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
-import { styles } from '../../styles/scenes/CryptoExchangeQuoteProcessingSceneStyles.js'
+import { THEME } from '../../theme/variables/airbitz.js'
+import { scale } from '../../util/scaling.js'
 import { SceneWrapper } from '../common/SceneWrapper.js'
 
 export function CryptoExchangeQuoteProcessingScreenComponent(props: {}) {
@@ -23,3 +24,28 @@ export function CryptoExchangeQuoteProcessingScreenComponent(props: {}) {
     </SceneWrapper>
   )
 }
+
+const rawStyles = {
+  top: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'space-around'
+  },
+  bottom: {
+    flex: 1
+  },
+  momentText: {
+    color: THEME.COLORS.WHITE,
+    width: '100%',
+    textAlign: 'center',
+    fontSize: scale(18),
+    marginBottom: scale(20)
+  },
+  findingText: {
+    color: THEME.COLORS.WHITE,
+    width: '100%',
+    textAlign: 'center',
+    fontSize: scale(14)
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
