+++ src/modules/UI/components/ExchangeQuote/ExchangeQuoteComponent.js
@@ -2,14 +2,14 @@
 
 import { createSimpleConfirmModal } from 'edge-components'
 import React, { Component } from 'react'
-import { Image, Platform, TouchableOpacity, View } from 'react-native'
+import { Image, Platform, StyleSheet, TouchableOpacity, View } from 'react-native'
 import IonIcon from 'react-native-vector-icons/Ionicons'
 
 import { launchModal } from '../../../../components/common/ModalProvider.js'
 import { EXCLAMATION, MATERIAL_COMMUNITY } from '../../../../constants/indexConstants'
 import s from '../../../../locales/strings.js'
-import { styles as sceneStyles } from '../../../../styles/scenes/CryptoExchangeQuoteSceneStyles.js'
 import { THEME } from '../../../../theme/variables/airbitz.js'
+import { scale } from '../../../../util/scaling.js'
 import FormattedText from '../../components/FormattedText/FormattedText.ui.js'
 import { Icon } from '../Icon/Icon.ui'
 
@@ -28,9 +28,8 @@ type Props = {
 }
 type State = {}
 
-class ExchangeQuoteComponent extends Component<Props, State> {
+export class ExchangeQuoteComponent extends Component<Props, State> {
   renderBottom = () => {
-    const styles = sceneStyles.quoteDetailContainer
     if (this.props.isTop) {
       return (
         <View style={styles.bottomRow}>
@@ -59,7 +58,6 @@ class ExchangeQuoteComponent extends Component<Props, State> {
   }
 
   renderHeadline = () => {
-    const styles = sceneStyles.quoteDetailContainer
     if (this.props.isEstimate) {
       return (
         <TouchableOpacity style={styles.headlineRow} onPress={this.showExplanationForEstimate}>
@@ -78,7 +76,6 @@ class ExchangeQuoteComponent extends Component<Props, State> {
   }
 
   render() {
-    const styles = sceneStyles.quoteDetailContainer
     const container = this.props.isTop ? styles.containerExpanded : styles.containerCollapsed
     return (
       <View style={styles.container}>
@@ -114,4 +111,126 @@ class ExchangeQuoteComponent extends Component<Props, State> {
   }
 }
 
-export { ExchangeQuoteComponent }
+const rawStyles = {
+  container: {
+    width: '90%',
+    borderRadius: 3
+  },
+  containerCollapsed: {
+    width: '100%',
+    minHeight: scale(65),
+    backgroundColor: THEME.COLORS.OPACITY_WHITE,
+    borderRadius: 3
+  },
+  containerExpanded: {
+    width: '100%',
+    height: scale(94),
+    backgroundColor: THEME.COLORS.OPACITY_WHITE,
+    borderRadius: 3
+  },
+  headlineEstimateText: {
+    color: THEME.COLORS.ACCENT_ORANGE,
+    fontSize: scale(14),
+    marginRight: THEME.rem(0.375)
+  },
+  headlineRow: {
+    flexDirection: 'row',
+    paddingTop: scale(10),
+    paddingBottom: scale(10),
+    alignItems: 'center'
+  },
+  headlineText: {
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(14),
+    paddingTop: scale(10),
+    paddingBottom: scale(10)
+  },
+  minerFeeText: {
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(14),
+    paddingTop: scale(5),
+    paddingBottom: scale(10)
+  },
+  minerFeeRightText: {
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(14),
+    paddingTop: scale(5),
+    paddingBottom: scale(10),
+    alignSelf: 'flex-end'
+  },
+  currencyNameText: {
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(16),
+    paddingTop: scale(10),
+    paddingLeft: scale(10)
+  },
+  cryptoAmountText: {
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(16),
+    paddingTop: scale(10),
+    paddingRight: scale(10),
+    textAlign: 'right'
+  },
+  walletNameText: {
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(16),
+    paddingLeft: scale(10),
+    paddingBottom: scale(9)
+  },
+  fiatAmountText: {
+    color: THEME.COLORS.WHITE,
+    fontSize: scale(16),
+    paddingRight: 10,
+    paddingBottom: scale(9),
+    textAlign: 'right'
+  },
+  topRow: {
+    flexDirection: 'row'
+  },
+  bottomRow: {
+    height: scale(45),
+    paddingRight: 10,
+    paddingLeft: 10
+  },
+  bottomContentBox: {
+    borderTopWidth: 1,
+    flexDirection: 'row',
+    borderColor: THEME.COLORS.WHITE
+  },
+  bottomContentBoxLeft: {
+    flex: 1
+  },
+  bottomContentBoxRight: {
+    flex: 1,
+    justifyContent: 'flex-end'
+  },
+  walletInfoContainer: {
+    flex: 8,
+    alignItems: 'flex-start'
+  },
+  amountInfoContainer: {
+    flex: 11,
+    alignItems: 'flex-end'
+  },
+  logoContainer: {
+    flex: 2,
+    alignItems: 'center',
+    paddingTop: scale(10),
+    paddingLeft: 5
+  },
+  iconContainer: {
+    position: 'relative',
+    height: scale(29),
+    width: scale(29),
+    backgroundColor: THEME.COLORS.OPACITY_WHITE,
+    borderRadius: 15,
+    alignItems: 'center',
+    justifyContent: 'space-around'
+  },
+  currencyIcon: {
+    height: scale(25),
+    width: scale(25),
+    resizeMode: 'contain'
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
